+++ Core/Csv/CsvSerializer.cs
@@ -1,13 +1,10 @@
 ﻿namespace Core.Csv
 {
 
-    /// <summary>
-    /// Serialize an object to csv file
-    /// </summary>
+
     public class CsvSerializer
     {
 
     }
-
 
 }
