+++ Core/Persist/IPersistentService.cs
@@ -1,6 +1,6 @@
 ﻿using Core.Service;
 
-namespace Core.DB
+namespace Core.Persist
 {
 
     public interface IPersistentService : IService
